+++ src/http/ClientHandler.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "../event/Event.h"
 #include "HttpParser.h"
+#include "HttpResponse.h"
 #include "Router.h"
 
 class ClientHandler {
