+++ src/http/HttpParser.h
@@ -42,22 +42,6 @@ struct HttpRequest {
     std::string version_;
 };
 
-struct HttpResponse {
-    HttpCommonHeader header_;
-    int status_code_;
-    std::string message_;
-    HttpDate date_;
-    std::string location_;
-    std::string body_;
-    std::string version_;
-    static ssize_t send_response(int client_fd, HttpResponse& response);
-    void clear();
-    static HttpResponse render_default_error_page(int status_code);
-    static HttpResponse render_error(
-        int status_code, const std::map<int, ErrorPageDirective>& error_pages,
-        const ServerConfig& servers);
-};
-
 class HttpParser {
    public:
     static HttpRequest http_request_parse(const std::string& buffer);
