+++ src/http/Router.cpp
@@ -58,60 +58,6 @@ const LocationConfig& Router::find_location(const ServerConfig& server,
     return *location;
 }
 
-HttpResponse HttpResponse::render_default_error_page(int status_code) {
-    HttpResponse response;
-    std::stringstream ss;
-    ss << "<!DOCTYPE html>\n"
-       << "<html>\n"
-       << "<head>\n"
-       << "<title>" << status_code << "</title>\n"
-       << "</head>\n"
-       << "<body>" << HttpStatus::reason(status_code) << "</body>\n"
-       << "</html>\n";
-    response.body_ = ss.str();
-    response.header_.content_type_ = "text/html";
-    response.header_.content_length_ = response.body_.size();
-    return response;
-}
-
-HttpResponse HttpResponse::render_error(
-    int status_code, const std::map<int, ErrorPageDirective>& error_pages,
-    const ServerConfig& servers) {
-    HttpResponse response;
-    response.status_code_ = status_code;
-    response.message_ = HttpStatus::reason(status_code);
-    const std::map<int, ErrorPageDirective>::const_iterator error_page =
-        error_pages.find(status_code);
-    if (error_page == error_pages.end()) {
-        return render_default_error_page(status_code);
-    }
-
-    const LocationConfig& location =
-        Router::find_location(servers, error_page->second.target);
-    int fd;
-    for (std::vector<std::string>::const_iterator index_file =
-             location.getCommonConfig().index_files_.begin();
-         index_file != location.getCommonConfig().index_files_.end();
-         ++index_file) {
-        fd = open((error_page->second.target +
-                   location.getCommonConfig().root_ + *index_file)
-                      .c_str(),
-                  O_RDONLY);
-        if (fd == -1) continue;
-        break;
-    }
-    char buf[1024];  // NOLINT
-    std::string buffer;
-    while (true) {
-        ssize_t len = read(fd, buf, sizeof(buf));
-        if (len == -1) throw std::runtime_error("read() failed");
-        if (len == 0) break;
-        buffer.append(buf);
-    }
-    response.body_ = buffer;
-    return response;
-}
-
 HttpResponse Router::create_response(const HttpRequest& request) {  // NOLINT
     HttpResponse response;
     // 対象の server を見つける
